Move non-Qt (GUI) code to `modmesh.plot`
`modmesh.plot` is a new sub-package to hold code for non-GUI drawing Python code.
